Fixes #6 and fixes #7
When not using the true button class and instead styling the `a` tag directly, the color values don't really get overwritten.
To fix this, one needs to explicitly set the text-color for all child-elements of `a`.
This is caused by the typography plugin.

The typography plugin also produced the issue of not being able to overwrite no-underline as the original previous typography definition actually defines the decorator.
A quick fix was to use `!important` to create an `force-no-underline` style

Also moved from emerald to corporate theme for light mode.